[refactor #27] 패키지 계층형 구조로 변경
