+++ elastic/els.yml
@@ -1,5 +1,11 @@
 version: '3.8'
 services:
+  redis:
+    image: redis:latest
+    container_name: docker_redis
+    ports:
+      - 6379:6379
+    restart: always
   es:
     build:
       context: .
@@ -21,7 +27,7 @@ services:
     image: docker.elastic.co/kibana/kibana:7.9.1
     environment:
       SERVER_NAME: kibana
-      ELASTICSEARCH_HOSTS: http://es:9200
+      ELASTICSEARCH_HOSTS: http:Elasaticrhk  //es:9200
     ports:
       - 5601:5601
     # Elasticsearch Start Dependency
