+++ src/main/java/com/soeun/GiftFunding/config/CustomPrincipalHandshakeHandler.java
@@ -1,7 +1,5 @@
-package com.soeun.GiftFunding.domain.chat;
+package com.soeun.GiftFunding.config;
 
-import java.security.Principal;
-import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -9,6 +7,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.server.support.DefaultHandshakeHandler;
+
+import java.security.Principal;
+import java.util.Map;
 
 @Slf4j
 public class CustomPrincipalHandshakeHandler extends DefaultHandshakeHandler {
