+++ src/main/java/com/soeun/GiftFunding/security/OAuth/OAuthApiClient.java
@@ -1,7 +1,7 @@
-package com.soeun.GiftFunding.OAuth;
+package com.soeun.GiftFunding.security.OAuth;
 
-import com.soeun.GiftFunding.OAuth.dto.OAuthInfoResponse;
-import com.soeun.GiftFunding.OAuth.dto.OAuthLoginRequest;
+import com.soeun.GiftFunding.security.OAuth.dto.OAuthInfoResponse;
+import com.soeun.GiftFunding.security.OAuth.dto.OAuthLoginRequest;
 import com.soeun.GiftFunding.type.OAuthProvider;
 
 public interface OAuthApiClient {
